agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-adk>=0.3.2",
    "ag-ui-langgraph>=0.0.21",
    "aiosqlite>=0.21.0",
    "copilotkit>=0.1.72",
    "fastapi>=0.115.14",
    "google-adk>=1.18.0",
    "google-genai>=1.50.1",
    "langchain-anthropic>=1.1.0",
    "langchain-core>=1.0.5",
    "langchain-google-genai>=3.0.3",
    "langgraph>=1.0.3",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "mcp>=1.21.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
